Remove Perl::Critic::Document monkey patch.
As of Perl::Critic 1.119, this module works with an unpatched
Perl::Critic::Document. The problem with 1.118 and earler was that
method element_is_in_lexical_scope_after_statement_containing() would
go into an endless loop in some cases, due to finding parent of the
wrong element. This module monkey-patched in a correct version of the
method, but that is no longer needed.


git-svn-id: file:///Users/tom/Code/svn/repos/Perl-Critic-Policy-Variables-ProhibitUnusedVarsStricter/trunk@4114 f6fd3633-1913-47c2-bad4-372d75c948ec