Allow 'return ( my $x = 0 )'
This construction can be used to return a temporary value from an lvalue
subroutine. The case 'return my $x = 0' is equivalent, but was not
recognized as defining a variable. Now both are recognized and accepted
by default.

If you want to reject these, there is a new Boolean configuration item,
prohibit_returned_lexicals.