Survive variable declarations that declare nothing
What this is really about is the use of Object::InsideOut, which
declares its own syntax. In rare cases this causes PPI to spit out
PPI::Statement::Variable objects that declare nothing. The real-world
example is from JABRA/MetasploitExpress-Parser-0.02.tar.gz file
lib/MetasploitExpress/Parser/Host.pm, which contains the line
    my @state : Field : Arg(state) : Get(state);
In this case the last two occurrances of the string 'state' get parsed
as PPI::Statement::Variable objects that do not declare anything. In
principal the same could happen with 'my' or 'our'.