Refactor handling of derived PPI documents.
We now track the PPI::Element they are derived from and treat that
element as the parent of the derived PPI::Document.

The implementation wraps all calls to ->ppi() with code to record the
associated PPI::Element, and wraps all calls to ->parent() with code to
retrieve and return the associated PPI::Element if ->parent() returned
nothing.

An author test was added to make sure the wrapped subroutines stayed
wrapped.

The old code passed around @scope_of_record arguments, not very
consistently.